requirements.txt
Django>=3.2.4,<4.2
djangorestframework>=3.12.4,<3.15
psycopg2-binary==2.9.3
flake8>=5.0,<6.0
drf-spectacular>=0.25.0,<0.26.5